refactor collection into package
