+++ src/lenskit/data/collection/__init__.py
@@ -1,0 +1,8 @@
+"""
+Item list collections.
+"""
+
+from ._base import ItemListCollection
+from ._keys import GenericKey, UserIDKey
+
+__all__ = ["GenericKey", "UserIDKey", "ItemListCollection"]
